Use squeue instead of scontrol for slurm job status. HTCONDOR-104
When filling the cache in slurm_status.py, we want to query all jobs
for the current user. scontrol won't do that (it does all jobs for all
users or a single job). On a busy system, the global query can take
several minutes. squeue can query all jobs for a single user, which
should be faster.

The squeue query doesn't give us WorkerNode (which HTCondor ignores) or
ExitCode (which we already query with sacct for completed jobs).