Fix the pbs and slurm status scripts for Python 3.
The new pickle format is binary data that we need to convert to ascii
when writing to a text file.
open() with 'a+' appears to now place the file position indicator to the
end of the file before any reads or writes. Use 'r+' instead when we
want to read.