Stop making symlinks in ~/.blah_jobproxy_dir/. HTCONDOR-520
To allow proxy refresh to work when the blahp makes a modified copy of
the proxy on submit, it creates a symlink to the "live" proxy file under
~/.blah_jobproxy_dir/, named with the job id. These symlinks are
frequently not cleaned up, and it's hard to guarantee that they are
always cleaned up.

The name for the modified proxy file is always the name of the original
file plus one of several well-known suffixes. We already assume that the
proxy refresh command will use the same proxy filename as the original
submit command. So instead of finding the live proxy file via a symlink
under HOME, we can check for the existence of the filename we would have
created at submit time.