Naming output files causing overwriting. HTCONDOR-553
When TransferOutputFiles is given, and the blahp scripts copy/move the
files, the same filename is used as the destination for all
copying/moving. This causes most output files to be missing from the job
sandbox.
In function bls_setup_all_files() in blah_common_submit_functions.sh,
the variable xfileremap needs to be set on every loop iteration.